requirements.txt
flask
flask-bcrypt
flask-cors
google-genai
python-dotenv
mysql-connector-python
